added master password protection

+++ src/main.py
@@ -1,20 +1,28 @@
+import base64
+import os
+from os import path
 from cryptography.fernet import Fernet
+from cryptography.hazmat.primitives import hashes
+from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
 
-"""def write_key():
-    key = Fernet.generate_key()
-    with open("key.key", "wb") as key_file:
-        key_file.write(key)"""
+master_pwd = input("What is the master password: ")
 
+def generate_key():
+    if path.exists("key.key"):
+        with open("key.key", "rb") as key_file:
+            key = key_file.read()
+            if master_pwd in key.decode():
+                return key
+            else:
+                raise SystemExit("Invalid Master Password!")
 
-def load_key():
-    with open("key.key", "rb") as key_file:
-        return key_file.read()
+    else:
+        key = Fernet.generate_key() + master_pwd.encode()
+        with open("key.key", "wb") as key_file:
+            key_file.write(key)
+            return key
 
-
-# master_pwd = input("What is the master password: ")
-key = load_key()  # + master_pwd.encode()
-fer = Fernet(key)
-
+fer = Fernet(generate_key())
 
 def view():
     with open("passwords.txt", "r") as f:
